requirements.txt
Flask
Pillow
mysql-connector
nose
pep8
autopep8